requirements.txt
ffmpeg-python
sounddevice
pyradios
pyqt5==5.15.2
requests
python-vlc
mplayer.py
qdarkstlye